Refactor ListPage component: Split into CharacterFilters and CharacterCard components
